VIDSOL-349 add strict mode init in debug
