+++ app/src/main/java/com/vonage/android/MainApplication.kt
@@ -2,8 +2,6 @@ package com.vonage.android
 
 import android.app.Application
 import com.vonage.android.notifications.VeraNotificationChannelRegistry
-import android.net.TrafficStats
-import android.os.StrictMode
 import dagger.hilt.android.HiltAndroidApp
 import javax.inject.Inject
 
@@ -17,27 +15,5 @@ open class MainApplication : Application() {
         super.onCreate()
 
         notificationChannelRegistry.createNotificationChannels()
-        if (BuildConfig.DEBUG) { // could be moved to debug source set
-            enableStrictMode()
-        }
-    }
-
-    private fun enableStrictMode() {
-        TrafficStats.setThreadStatsTag(APP_THREAD_TAG)
-        StrictMode.setThreadPolicy(
-            StrictMode.ThreadPolicy.Builder()
-                .detectAll()
-                .permitDiskReads()
-                .build()
-        )
-        StrictMode.setVmPolicy(
-            StrictMode.VmPolicy.Builder()
-                .detectAll()
-                .build()
-        )
-    }
-
-    private companion object {
-        const val APP_THREAD_TAG = 10000
     }
 }
